+++ src/codegen.c
@@ -690,7 +690,7 @@ static mrb_sym
 attrsym(codegen_scope *s, mrb_sym a)
 {
   const char *name;
-  int len;
+  size_t len;
   char *name2;
 
   name = mrb_sym2name_len(s->mrb, a, &len);
@@ -789,7 +789,7 @@ gen_call(codegen_scope *s, node *tree, mrb_sym name, int sp, int val)
   }
   pop_n(n+1);
   {
-    int len;
+    size_t len;
     const char *name = mrb_sym2name_len(s->mrb, sym, &len);
 
     if (!noop && len == 1 && name[0] == '+')  {
@@ -1491,7 +1491,7 @@ codegen(codegen_scope *s, node *tree, int val)
   case NODE_OP_ASGN:
     {
       mrb_sym sym = sym(tree->cdr->car);
-      int len;
+      size_t len;
       const char *name = mrb_sym2name_len(s->mrb, sym, &len);
       int idx;
